requirements.txt
# Core dependencies
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
questionary>=1.10.0
rich>=13.0.0
typer>=0.9.0
python-dotenv>=1.0.0

# Optional: For better performance
tiktoken>=0.5.0
blobfile>=3.0.0
